requirements.txt
plac >= 1.2
PyYAML >= 5.3
